refactor(🎨 Themes): Refactored themes folder structure
