+++ src/index.ts
@@ -2,5 +2,4 @@ export * from './BlueBase';
 export { BlueBaseApp } from './components/';
 export * from './Context';
 export * from './getComponent';
-export * from './models';
 export * from './registries';
