--- src/models/Theme/index.ts
@@ -1,3 +1,0 @@
-export * from './applyStyles';
-export * from './buildTheme';
-export * from './Theme';
