--- src/models/index.ts
@@ -1,2 +1,0 @@
-export * from './Configs';
-export * from './Theme';
