+++ src/themes/index.ts
@@ -1,3 +1,4 @@
-export * from './BlueBaseDarkTheme';
-export * from './BlueBaseLightTheme';
+export * from './helpers';
+export * from './structure';
+export * from './themes';
 export * from './ThemeContext';
