+++ src/themes/structure/index.ts
@@ -1,0 +1,3 @@
+export * from './Palette';
+export * from './Theme';
+export * from './Typography';
