+++ src/themes/themes/BlueBaseLightTheme.ts
@@ -1,7 +1,7 @@
-import { ThemeInput } from '../registries';
-import { createPalette } from './createPalette';
-import { createTheme } from './createTheme';
-import { createTypography } from './createTypography';
+import { ThemeInput } from '../../registries';
+import { createPalette } from '../helpers/createPalette';
+import { createTheme } from '../helpers/createTheme';
+import { createTypography } from '../helpers/createTypography';
 
 const palette = createPalette('light');
 const typography = createTypography(palette);
