+++ src/themes/themes/index.ts
@@ -1,0 +1,2 @@
+export * from './BlueBaseDarkTheme';
+export * from './BlueBaseLightTheme';
